+++ api/BUILD.bazel
@@ -21,6 +21,8 @@ py_library(
         "analysis_views.py",
         "apps.py",
         "auth_account_views.py",
+        "auth_export_archive.py",
+        "auth_export_data.py",
         "auth_export_views.py",
         "dev_bootstrap.py",
         "auth_google_views.py",
@@ -115,6 +117,8 @@ filegroup(
         "analysis_views.py",
         "apps.py",
         "auth_account_views.py",
+        "auth_export_archive.py",
+        "auth_export_data.py",
         "auth_export_views.py",
         "dev_bootstrap.py",
         "auth_google_views.py",
